added purchase controller and purchase related functions at user and store controllers

+++ CommerceSystem/src/Server/Domain/ExternalComponents/PaymentSystem.java
@@ -9,7 +9,7 @@ public class PaymentSystem {
     Actual payment from the external system, requires amount of cash to spend from
     the given bankAccount
      */
-    public boolean pay (int amount, int bankAccount){
+    public boolean pay (double amount, int bankAccount){
         return true; /* as to version 1, this payment will always work */
     }
 }
